Extract serialization helpers
